new structure for move_mir_tasks package
