+++ move_mir_tasks/src/solution_01.cpp
