--- move_mir_tasks/task_02/README.md
@@ -1,9 +1,0 @@
-# Task 2
-
-Now let the robot rotate 180 degrees and move back to the starting point in C++ or Python
-
-### Hints
-- Use the following activity diagram to code a state machine
-
-![Activity diagram](https://i.ibb.co/cDJ16p1/Diagram-2022-12-02-13-56-49.png "Activity diagram")
-- For the orientation you need to work with quaternions. Use the [TF2 tutorial](http://wiki.ros.org/tf2/Tutorials/Quaternions "TF2 tutorial") for that
