--- move_mir_tasks/task_03/README.md
@@ -1,6 +1,0 @@
-# Task 2
-
-Write a loop for each of the following pose topics that completes the tasks 01 and 02 an infinite number of times:
-1. Gazebo ground_truth
-2. AMCL 
-3. Odometry
