--- move_mir_tasks/task_03/solution_node3/launch/solution_03.launch
@@ -1,9 +1,0 @@
-<launch>
-	<include file="$(find match_gazebo)/launch/big_square.launch"/>
-
-	<include file="$(find mir_launch_sim)/launch/mir_100.launch"/>
-	
-	<node type="rviz" name="rviz" pkg="rviz" args="-d $(find mir_examples)/rviz/single_mir_visualization.rviz" /><!--2D Navigation-->
-
-</launch>
-
